Set the top and left margins of the sidebar.
With bootstrap, space between the sidebar and the left and top edges of
the screen disappeared. The top margin changed depending on the content
of the body (e.g. Camping versus Nature Info -> Plants). By defining the
margins of the body in style.css, the sidebar no longer moves when
clicking the sidebar links.