Allow submitting a camping reservation more than once.
Also handles form submission with the ng-submit attribute of the form
object instead of the ng-click attribute of the submit button. This allows the
user to submit the form using any its input elements; for example, now
the user can press the enter key in the From field to submit the form.