Fix the inconsistency between the switch state (animal or plant) and the form.
Also changed the variable name "sight" to "type" to indicate that this
is the type of sighting being reported (Animal or Plant). Changed the type
dependent part of the form to compare the type against the Types
"enum"; this can make it simpler to add more types in the future
and decouples this condition from the type of the variable. This
decoupling is useful, for one, it allowed me to redefine the animal
type as false so that the "off" position of the switch corresponds
to the animal type and the label without needing to change the
form's ng-show condition.